# docs/reference/api/openapi.yaml
openapi: 3.1.0
info:
  title: Capacitarr API
  version: 3.0.0
  description: |
    Capacitarr is a capacity management tool for media servers. It monitors disk
    usage across Sonarr, Radarr, Lidarr, Readarr, Plex, Jellyfin, Emby, Tautulli,
    and Seerr integrations, scores media items for potential cleanup based on
    configurable weights and rules, and can automatically (or via approval) free
    disk space when thresholds are breached.

    ## Authentication

    Most endpoints require authentication via one of:
    - **JWT cookie** (`jwt`) — set automatically on login
    - **Bearer token** — `Authorization: Bearer <token>` header
    - **API key** — `X-Api-Key: <key>` header

    The `/health`, `/version`, `/auth/login`, and `/auth/logout` endpoints are public.
  license:
    name: PolyForm Noncommercial 1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
  - url: http://localhost:2187/api/v1
    description: Local development server
security:
  - cookieAuth: []
  - bearerAuth: []
  - apiKey: []
...
paths:
...
  /integrations:
    get:
      operationId: listIntegrations
      tags: [Integrations]
      summary: List all integrations
      description: Returns all configured integrations. API keys are masked in the response (first 4 + last 4 characters shown).
      responses:
        "200":
          description: List of integrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IntegrationConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      operationId: createIntegration
      tags: [Integrations]
      summary: Create a new integration
      description: |
        Creates a new integration configuration. The `type`, `name`, `url`, and
        `apiKey` fields are required. The URL must use HTTP or HTTPS scheme.
        Valid types: `plex`, `sonarr`, `radarr`, `lidarr`, `readarr`, `tautulli`,
        `seerr`, `jellyfin`, `emby`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationConfigCreate"
      responses:
        "201":
          description: Integration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /integrations/{id}:
    get:
      operationId: getIntegration
      tags: [Integrations]
      summary: Get a single integration
      description: Returns a single integration by ID. API key is masked.
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Integration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Integration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateIntegration
      tags: [Integrations]
      summary: Update an integration
      description: |
        Updates an existing integration. Only provided fields are updated.
        If the `apiKey` field contains a masked value (with `...`), it is
        ignored and the existing key is preserved.
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegrationConfigUpdate"
      responses:
        "200":
          description: Updated integration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Integration not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteIntegration
      tags: [Integrations]
      summary: Delete an integration
      description: Permanently deletes an integration configuration.
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Integration deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /integrations/test:
    post:
      operationId: testIntegrationConnection
      tags: [Integrations]
      summary: Test integration connection
      description: |
        Tests connectivity to a service using the provided credentials.
        Returns `success: true` on success, or `success: false` with an
        error message on failure. Note: the HTTP status is always 200 —
        the `success` field indicates the test result.
        Optionally pass `integrationId` to invalidate the rule value cache
        for that integration on successful test.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionTestRequest"
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionTestResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /integrations/sync:
    post:
      operationId: syncIntegrations
      tags: [Integrations]
      summary: Sync all integrations
      description: |
        Triggers a manual synchronization of all enabled integrations.
        Tests each connection, fetches disk space and media item counts,
        and updates disk groups. Invalidates all rule value caches.
      responses:
        "200":
          description: Sync results for each integration
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/SyncResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ===========================================================================
  # Preferences
  # ===========================================================================
...
components:
...
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key generated via `POST /auth/apikey`
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login response
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: HttpOnly JWT cookie set on login

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: integer

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request — invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
      required: [error]
      example:
        error: "Invalid request body"
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
      example:
        message: "Password changed successfully"

    # ---- Health / Version ----
...
    IntegrationConfig:
      type: object
      properties:
        id:
          type: integer
          example: 1
        type:
          type: string
          enum: [plex, sonarr, radarr, lidarr, readarr, tautulli, seerr, jellyfin, emby, tracearr, jellystat]
          example: sonarr
        name:
          type: string
          example: "Sonarr (TV)"
        url:
          type: string
          format: uri
          example: "http://sonarr:8989"
        apiKey:
          type: string
          description: Masked API key (first 4 + last 4 characters)
          example: "abcd...wxyz"
        enabled:
          type: boolean
          example: true
        collectionDeletion:
          type: boolean
          description: When enabled, deleting one collection member deletes all
          default: false
          example: false
        mediaSizeBytes:
          type: integer
          format: int64
          example: 549755813888
        mediaCount:
          type: integer
          example: 150
        lastSync:
          type: ["string", "null"]
          format: date-time
        lastError:
          type: string
          example: ""
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, type, name, url, apiKey, enabled]
    IntegrationConfigCreate:
      type: object
      properties:
        type:
          type: string
          enum: [plex, sonarr, radarr, lidarr, readarr, tautulli, seerr, jellyfin, emby, tracearr, jellystat]
        name:
          type: string
        url:
          type: string
          format: uri
          description: Must use HTTP or HTTPS scheme
        apiKey:
          type: string
          description: Plaintext API key or token for the service
      required: [type, name, url, apiKey]
    IntegrationConfigUpdate:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        apiKey:
          type: string
          description: New API key, or masked value to keep existing
        enabled:
          type: boolean
    ConnectionTestRequest:
      type: object
      properties:
        type:
          type: string
          enum: [plex, sonarr, radarr, lidarr, readarr, tautulli, seerr, jellyfin, emby, tracearr, jellystat]
        url:
          type: string
          format: uri
        apiKey:
          type: string
        integrationId:
          type: ["integer", "null"]
          description: Optional — invalidates rule value cache for this integration on success
      required: [type, url, apiKey]
    ConnectionTestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          description: Present on success
        error:
          type: string
          description: Present on failure
      required: [success]
      example:
        success: true
        message: "Connection successful"
    SyncResult:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [ok, error]
        error:
          type: string
          description: Present if status is error
        diskError:
          type: string
        diskSpace:
          type: array
          items:
            $ref: "#/components/schemas/DiskSpace"
        mediaError:
          type: string
        mediaCount:
          type: integer
      required: [id, name, type, status]
    DiskSpace:
      type: object
      properties:
        path:
          type: string
          example: "/media"
        totalBytes:
          type: integer
          format: int64
        freeBytes:
          type: integer
          format: int64
      required: [path, totalBytes, freeBytes]

    # ---- Preferences ----
...